Fix tree status updates not reflecting in UI (#457)
* Build Dashboard parent widget

* random injector
